# openapi-codegen/docs/openapi/test.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sample CRUD API",
    "version": "1.0.0"
  },
  "paths": {
    "/items": {
      "get": {
        "summary": "Get items with filter",
        "operationId": "getItems",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "Filter by item name",
            "required": false,
            "schema": {
              "type": "string",
              "default": "Book"
            },
            "example": "Book"
          },
           {
            "name": "name1",
            "in": "query",
            "description": "Filter by item name",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "Book"
          }
        ],
        "responses": {
          "200": {
            "description": "List of items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Item"
                  }
                },
                "examples": {
                  "example": {
                    "value": [
                      {
                        "id": 1,
                        "name": "Book",
                        "price": 9.99
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create an item",
        "operationId": "createItem",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemCreate"
              },
              "examples": {
                "example": {
                  "value": {
                    "name": "Pen",
                    "price": 1.5
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Item created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                },
                "examples": {
                  "example": {
                    "value": {
                      "id": 2,
                      "name": "Pen",
                      "price": 1.5
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/items/{id}": {
      "get": {
        "summary": "Get item by ID",
        "operationId": "getItemById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Item details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                },
                "examples": {
                  "example": {
                    "value": {
                      "id": 1,
                      "name": "Book",
                      "price": 9.99
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update an item",
        "operationId": "updateItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemCreate"
              },
              "examples": {
                "example": {
                  "value": {
                    "name": "Notebook",
                    "price": 3.25
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete an item",
        "operationId": "deleteItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Item deleted"
          }
        }
      },
      "patch": {
        "summary": "Partially update an item",
        "operationId": "patchItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "price": { "type": "number" }
                }
              },
              "examples": {
                "example": {
                  "value": {
                    "price": 4.5
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Item": {
        "type": "object",
        "required": ["id", "name", "price"],
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "price": { "type": "number" }
        }
      },
      "ItemCreate": {
        "type": "object",
        "required": ["name", "price"],
        "properties": {
          "name": { "type": "string" },
          "price": { "type": "number" }
        }
      }
    }
  }
}